+++ product-service/src/main/java/com/oingmaryho/business/productservice/presentation/dto/response/ProductSearchResponseDto.java
@@ -11,7 +11,7 @@ public record ProductSearchResponseDto(
 	UUID manageHubId,
 	UUID companyId,
 	String companyName,
-	Long price,
-	Long stock
+	Integer price,
+	Integer stock
 ) {
 }
